+++ packages/backend/src/core/IdService.ts
@@ -8,6 +8,7 @@ import { ulid } from 'ulid';
 import { DI } from '@/di-symbols.js';
 import type { Config } from '@/config.js';
 import { genAid, parseAid } from '@/misc/id/aid.js';
+import { genAidx, parseAidx } from '@/misc/id/aidx.js';
 import { genMeid, parseMeid } from '@/misc/id/meid.js';
 import { genMeidg, parseMeidg } from '@/misc/id/meidg.js';
 import { genObjectId, parseObjectId } from '@/misc/id/object-id.js';
@@ -31,6 +32,7 @@ export class IdService {
 
 		switch (this.method) {
 			case 'aid': return genAid(date);
+			case 'aidx': return genAidx(date);
 			case 'meid': return genMeid(date);
 			case 'meidg': return genMeidg(date);
 			case 'ulid': return ulid(date.getTime());
@@ -43,6 +45,7 @@ export class IdService {
 	public parse(id: string): { date: Date; } {
 		switch (this.method) {
 			case 'aid': return parseAid(id);
+			case 'aidx': return parseAidx(id);
 			case 'objectid': return parseObjectId(id);
 			case 'meid': return parseMeid(id);
 			case 'meidg': return parseMeidg(id);